Add "umask" option
Adds the option "umask" to the "daemonize" function to control what the
umask is set to while starting the daemon. It was previously always set
to 0 (this remains the default for compatibility), causing the PID,
STDOUT and STDERR files to have 666 permissions. A new test for this
option is included.